@@ -1,32 +1,29 @@
 name: craic_chat
 description: A new Flutter application.
 
-
 version: 1.0.0+1
 
 environment:
-  sdk: ">=2.7.0 <3.0.0"
+  sdk: ">=2.12.0 <4.0.0"
 
 dependencies:
   flutter:
     sdk: flutter
 
-
-  cupertino_icons: ^0.1.2
-  animated_text_kit: ^3.1.2
-  firebase_core: ^1.0.1
-  firebase_auth: ^1.0.1
-  cloud_firestore: ^1.0.1
-  modal_progress_hud: ^0.1.3
+  cupertino_icons: ^1.0.2
+  animated_text_kit: ^4.2.2
+  firebase_core: ^2.24.2
+  firebase_auth: ^4.15.2
+  cloud_firestore: ^4.13.6
+  modal_progress_hud_nsn: ^0.5.1
 
 dev_dependencies:
   flutter_test:
     sdk: flutter
+  flutter_lints: ^3.0.1
 
 flutter:
-
   uses-material-design: true
 
   assets:
-  - images/
-
+    - images/
