@@ -3,20 +3,22 @@ import 'package:flutter/material.dart';
 import 'package:craic_chat/components/rounded_button.dart';
 import 'package:craic_chat/constants.dart';
 import 'package:firebase_auth/firebase_auth.dart';
-import 'package:modal_progress_hud/modal_progress_hud.dart';
+import 'package:modal_progress_hud_nsn/modal_progress_hud_nsn.dart';
 
 class LoginScreen extends StatefulWidget {
-  static String id = 'login_screen';
+  static const String id = 'login_screen';
 
   @override
-  _LoginScreenState createState() => _LoginScreenState();
+  State<LoginScreen> createState() => _LoginScreenState();
 }
 
 class _LoginScreenState extends State<LoginScreen> {
-  final _auth = FirebaseAuth.instance;
+  final FirebaseAuth _auth = FirebaseAuth.instance;
   bool showSpinner = false;
-  String email;
-  String password;
+
+  // FIX 1: null-safety
+  late String email;
+  late String password;
 
   @override
   Widget build(BuildContext context) {
@@ -25,7 +27,7 @@ class _LoginScreenState extends State<LoginScreen> {
       body: ModalProgressHUD(
         inAsyncCall: showSpinner,
         child: Padding(
-          padding: EdgeInsets.symmetric(horizontal: 24.0),
+          padding: const EdgeInsets.symmetric(horizontal: 24.0),
           child: Column(
             mainAxisAlignment: MainAxisAlignment.center,
             crossAxisAlignment: CrossAxisAlignment.stretch,
@@ -33,27 +35,24 @@ class _LoginScreenState extends State<LoginScreen> {
               Flexible(
                 child: Hero(
                   tag: 'logo',
-                  child: Container(
+                  child: SizedBox(
                     height: 200.0,
                     child: Image.asset('images/logo_non.png'),
                   ),
                 ),
               ),
-              SizedBox(
-                height: 48.0,
-              ),
+              const SizedBox(height: 48.0),
               TextField(
                 keyboardType: TextInputType.emailAddress,
                 textAlign: TextAlign.center,
                 onChanged: (value) {
                   email = value;
                 },
-                decoration:
-                kTextFieldDecoration.copyWith(hintText: 'Enter your email'),
+                decoration: kTextFieldDecoration.copyWith(
+                  hintText: 'Enter your email',
+                ),
               ),
-              SizedBox(
-                height: 8.0,
-              ),
+              const SizedBox(height: 8.0),
               TextField(
                 obscureText: true,
                 textAlign: TextAlign.center,
@@ -61,11 +60,10 @@ class _LoginScreenState extends State<LoginScreen> {
                   password = value;
                 },
                 decoration: kTextFieldDecoration.copyWith(
-                    hintText: 'Enter your password'),
+                  hintText: 'Enter your password',
+                ),
               ),
-              SizedBox(
-                height: 24.0,
-              ),
+              const SizedBox(height: 24.0),
               RoundedButton(
                 title: 'Log In',
                 color: Colors.lightBlueAccent,
@@ -73,20 +71,24 @@ class _LoginScreenState extends State<LoginScreen> {
                   setState(() {
                     showSpinner = true;
                   });
-                  try {
-                    final user = _auth.signInWithEmailAndPassword(
-                        email: email, password: password);
 
-                    if (user != null) {
+                  try {
+                    // FIX 2: await is REQUIRED
+                    final userCredential =
+                        await _auth.signInWithEmailAndPassword(
+                      email: email,
+                      password: password,
+                    );
+
+                    if (userCredential.user != null) {
                       Navigator.pushNamed(context, ChatScreen.id);
                     }
-
+                  } on FirebaseAuthException catch (e) {
+                    print(e.message);
+                  } finally {
                     setState(() {
                       showSpinner = false;
                     });
-                  }
-                  catch (e){
-                    print(e);
                   }
                 },
               ),
