@@ -3,20 +3,22 @@ import 'package:craic_chat/screens/chat_screen.dart';
 import 'package:flutter/material.dart';
 import 'package:craic_chat/components/rounded_button.dart';
 import 'package:firebase_auth/firebase_auth.dart';
-import 'package:modal_progress_hud/modal_progress_hud.dart';
+import 'package:modal_progress_hud_nsn/modal_progress_hud_nsn.dart';
 
 class RegistrationScreen extends StatefulWidget {
-  static String id = 'registration_screen';
+  static const String id = 'registration_screen';
 
   @override
-  _RegistrationScreenState createState() => _RegistrationScreenState();
+  State<RegistrationScreen> createState() => _RegistrationScreenState();
 }
 
 class _RegistrationScreenState extends State<RegistrationScreen> {
-  final _auth = FirebaseAuth.instance;
+  final FirebaseAuth _auth = FirebaseAuth.instance;
   bool showSpinner = false;
-  String email;
-  String password;
+
+  // Null-safety fix
+  late String email;
+  late String password;
 
   @override
   Widget build(BuildContext context) {
@@ -25,7 +27,7 @@ class _RegistrationScreenState extends State<RegistrationScreen> {
       body: ModalProgressHUD(
         inAsyncCall: showSpinner,
         child: Padding(
-          padding: EdgeInsets.symmetric(horizontal: 24.0),
+          padding: const EdgeInsets.symmetric(horizontal: 24.0),
           child: Column(
             mainAxisAlignment: MainAxisAlignment.center,
             crossAxisAlignment: CrossAxisAlignment.stretch,
@@ -33,27 +35,24 @@ class _RegistrationScreenState extends State<RegistrationScreen> {
               Flexible(
                 child: Hero(
                   tag: 'logo',
-                  child: Container(
+                  child: SizedBox(
                     height: 200.0,
                     child: Image.asset('images/logo_non.png'),
                   ),
                 ),
               ),
-              SizedBox(
-                height: 48.0,
-              ),
+              const SizedBox(height: 48.0),
               TextField(
                 keyboardType: TextInputType.emailAddress,
                 textAlign: TextAlign.center,
                 onChanged: (value) {
                   email = value;
                 },
-                decoration:
-                    kTextFieldDecoration.copyWith(hintText: 'Enter your email'),
+                decoration: kTextFieldDecoration.copyWith(
+                  hintText: 'Enter your email',
+                ),
               ),
-              SizedBox(
-                height: 8.0,
-              ),
+              const SizedBox(height: 8.0),
               TextField(
                 obscureText: true,
                 textAlign: TextAlign.center,
@@ -61,32 +60,34 @@ class _RegistrationScreenState extends State<RegistrationScreen> {
                   password = value;
                 },
                 decoration: kTextFieldDecoration.copyWith(
-                    hintText: 'Enter your password'),
+                  hintText: 'Enter your password',
+                ),
               ),
-              SizedBox(
-                height: 24.0,
-              ),
+              const SizedBox(height: 24.0),
               RoundedButton(
-                title: 'Registration',
+                title: 'Register',
                 color: Colors.blueAccent,
                 onPressed: () async {
                   setState(() {
                     showSpinner = true;
                   });
+
                   try {
-                    final newUser = await _auth.createUserWithEmailAndPassword(
-                        email: email, password: password);
-                    if (newUser != null) {
+                    final userCredential =
+                        await _auth.createUserWithEmailAndPassword(
+                      email: email,
+                      password: password,
+                    );
+
+                    if (userCredential.user != null) {
                       Navigator.pushNamed(context, ChatScreen.id);
                     }
-
-                    setState(
-                      () {
-                        showSpinner = false;
-                      },
-                    );
-                  } catch (e) {
-                    print(e);
+                  } on FirebaseAuthException catch (e) {
+                    print(e.message);
+                  } finally {
+                    setState(() {
+                      showSpinner = false;
+                    });
                   }
                 },
               ),
